made the following changes to the code: 1. Visualization of the Graspability Score by subscribing to the topic /graspability_score. 2. Full ROS2 integration.
