--- detect_graspable_points/launch/detect_graspable_points.launch
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<launch>
-
-
-  <!-- This description means C++ codes related to graspable points detection -->
-  <node name="detect_graspable_points_main" pkg="detect_graspable_points" type="detect_graspable_points_main" output="screen" >
-  
-    <!-- <param name="use_simulation" value="$(arg use_simulation)" launch-prefix="valgrind" /> -->
-  </node>
-
-  <!-- <node pkg="tf2_ros" type="static_transform_publisher" name="camera_link_broadcaster" 
-    args="0 0 0 0 0 1.570796 camera_link camera_depth_optical_frame " /> -->
-
-
-</launch>
