+++ detect_graspable_points/launch/detect_graspable_points_launch.py
@@ -1,0 +1,12 @@
+from launch import LaunchDescription
+from launch_ros.actions import Node
+
+def generate_launch_description():
+    return LaunchDescription([
+        Node(
+            package='detect_graspable_points',
+            executable='detect_graspable_points_main',
+            name='detect_graspable_points_main',
+            output='screen'
+        )
+    ])
