--- detect_graspable_points/launch/rviz.launch
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<launch>
-
-
-  
-  <node pkg="image_publisher" type="image_publisher" name="image_publisher" args="$(find detect_graspable_points)/RVIZ/Legende_5.png" >
-  </node>
-  <node type="rviz" name="rviz" pkg="rviz" args="-d $(find detect_graspable_points)/RVIZ/HubRobo_RVIZ_20231201.rviz" />
-  
-  <!-- <node pkg="tf2_ros" type="static_transform_publisher" name="camera_link_broadcaster" 
-    args="0 0 0 0 0 1.570796 camera_link camera_depth_optical_frame " /> -->
-
-
-</launch>
