+++ detect_graspable_points/package.xml
@@ -1,36 +1,26 @@
 <?xml version="1.0"?>
-<package format="2">
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
   <name>detect_graspable_points</name>
   <version>0.0.0</version>
-  <description>The detect_graspable_points package</description>
+  <description>TODO: Package description</description>
+  <maintainer email="jitao.zheng@tum.de">jitao</maintainer>
+  <license>TODO: License declaration</license>
 
-  <!-- One maintainer tag required, multiple allowed, one person per tag -->
-  <!-- Example:  -->
-  <!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-  <maintainer email="naomasa@todo.todo">naomasa</maintainer>
+  <buildtool_depend>ament_cmake</buildtool_depend>
 
-
-  <!-- One license tag required, multiple allowed, one license per tag -->
-  <!-- Commonly used license strings: -->
-  <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
-  <license>TODO</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <depend>roscpp</depend>
-  <depend>rospy</depend>
+  <depend>rclcpp</depend>
   <depend>std_msgs</depend>
-  <depend>tf</depend>
-  <depend>pcl_ros</depend>
   <depend>sensor_msgs</depend>
   <depend>pcl_conversions</depend>
+  <depend>pcl_ros</depend>
   <depend>visualization_msgs</depend>
+  <exec_depend>ros2launch</exec_depend>
+  
+  <test_depend>ament_lint_auto</test_depend>
+  <test_depend>ament_lint_common</test_depend>
 
-
-
-
-  <!-- The export tag contains other, unspecified, tags -->
   <export>
-    <!-- Other tools can request additional information be placed here -->
-
+    <build_type>ament_cmake</build_type>
   </export>
 </package>
