+++ detect_graspable_points/pcd/peaks_artificial_rocks_0.pcd
@@ -1,0 +1,25 @@
+# .PCD v0.7 - Point Cloud Data file format
+VERSION 0.7
+FIELDS x y z rgb
+SIZE 4 4 4 4
+TYPE F F F U
+COUNT 1 1 1 1
+WIDTH 14
+HEIGHT 1
+VIEWPOINT 0 0 0 1 0 0 0
+POINTS 14
+DATA ascii
+-1.7836502 -3.5144796 0.49704629 4278255360
+0.39079624 -1.0065584 0.094039351 4278255360
+0.38053268 -1.0018985 0.089248985 4278255360
+0.38197821 -0.98793006 0.094521135 4278255360
+0.37171489 -0.98327041 0.089730769 4278255360
+0.24035355 -0.90985274 0.087712973 4278255360
+0.22380453 -0.89756703 0.088151246 4278255360
+-0.41870654 -1.2098289 0.038936704 4278255360
+-0.39937884 -1.1939725 0.039045483 4278255360
+-0.34076661 0.58689785 0.074977785 4278255360
+-0.35801035 0.59214807 0.075279266 4278255360
+-0.35033542 0.59859324 0.070324153 4278255360
+-0.34905839 0.61286974 0.06060043 4278255360
+-0.35990408 0.61028862 0.06567052 4278255360
